Improve file handling and add WebDav tests
The paths returned by the local filesystem are now correctly formatted regardless of the system's preferred path separator. The MoveToAsync method now correctly handles when the destination is a directory or a file. This update also includes the addition of an option for write permissions in WebDav and the necessary tests in the Aiursoft.Static.Tests project.